requirements.txt
Pyrebase4
thread6
numpy
pandas
json
requests